Prediction fetch working
